requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
cloudinary==1.32.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
passlib==1.7.4
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.6.0
python-decouple==3.8
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
zxcvbn==4.4.28
